refactor(compiler): split control-flow/plan/types.ts by Plan kind
The 397-line `types.ts` aggregated four Plan families (insert, loop,
composite, event-delegation) plus their shared helpers in one file.
Split into dedicated files alongside the corresponding builders /
stringifiers, with `types.ts` retained as a barrel re-export so
existing imports keep working.

Layout after split:

  control-flow/plan/
  ├── common.ts             # ScopeRef, ReactiveEffectsPassthrough
  ├── insert.ts             # InsertPlan + arms + bindings
  ├── loop.ts               # Plain/Component/Composite/Static + NestedComponentInit
  ├── event-delegation.ts   # EventDelegationPlan + ItemLookup family
  └── types.ts              # barrel re-export (back-compat for existing callers)

No call-site changes needed — every existing
`import { ... } from '../plan/types'` continues to resolve via the
barrel. Future callers can target the specific file directly.

The pre-PR-2 migration breadcrumbs ("Until PR 2 lands…") have been
trimmed from the file headers; that history lives in the PR descriptions
and git log now.

Verified: jsx 766 / 766, adapter-tests 214 / 214, client 212 / 212;
clean `bun run --filter '@barefootjs/jsx' build` exits 0.
